clipos: fs: Support for O_MAYEXEC flag for sys_open()
Add support for a O_MAYEXEC flag on sys_open(). When this flag is
passed, sys_open will check that the mount options for the underlying
vfsmount do not prevent execution. This may allow an interpreter to
check exec permissions before reading commands from a file, and thus
enforce W^X at their level.

WARNING: The value of O_MAYEXEC is the same as __FMODE_EXEC since commit
55ec82176eca which appeared upstream in 2009. Due to the purpose and
current use of __FMODE_EXEC, this is not an issue for now but should
still be kept in mind.

Cherry-picked from CLIP OS 4 patch: 1901_open_mayexec.patch